router: refactor existing header formatter/parser into separate files
Only changes are to BUILD files and include directives. Code changes are
purely copy/paste.
