+++ source/common/router/header_parser.cc
@@ -1,12 +1,6 @@
-#include "common/router/req_header_formatter.h"
+#include "common/router/header_parser.h"
 
-#include <cstdint>
-#include <string>
-
-#include "common/config/rds_json.h"
 #include "common/protobuf/utility.h"
-
-#include "fmt/format.h"
 
 namespace Envoy {
 namespace Router {
@@ -51,25 +45,39 @@ void RequestHeaderParser::evaluateRequestHeaders(Http::HeaderMap& headers,
   }
 }
 
-RequestHeaderFormatter::RequestHeaderFormatter(const std::string& field_name, bool append)
-    : append_(append) {
-  if (field_name == "PROTOCOL") {
-    field_extractor_ = [](const Envoy::AccessLog::RequestInfo& request_info) {
-      return Envoy::AccessLog::AccessLogFormatUtils::protocolToString(request_info.protocol());
-    };
-  } else if (field_name == "CLIENT_IP") {
-    field_extractor_ = [](const Envoy::AccessLog::RequestInfo& request_info) {
-      return request_info.getDownstreamAddress();
-    };
-  } else {
-    throw EnvoyException(
-        fmt::format("field '{}' not supported as custom request header", field_name));
+ResponseHeaderParserPtr ResponseHeaderParser::parse(
+    const Protobuf::RepeatedPtrField<envoy::api::v2::HeaderValueOption>& headers_to_add,
+    const Protobuf::RepeatedPtrField<ProtobufTypes::String>& headers_to_remove) {
+  ResponseHeaderParserPtr response_header_parser(new ResponseHeaderParser());
+
+  for (const auto& header_value_option : headers_to_add) {
+    response_header_parser->headers_to_add_.emplace_back(
+        HeaderAddition{Http::LowerCaseString(header_value_option.header().key()),
+                       header_value_option.header().value(),
+                       PROTOBUF_GET_WRAPPED_OR_DEFAULT(header_value_option, append, true)});
   }
+
+  for (const std::string& header : headers_to_remove) {
+    response_header_parser->headers_to_remove_.emplace_back(Http::LowerCaseString(header));
+  }
+
+  return response_header_parser;
 }
 
-const std::string
-RequestHeaderFormatter::format(const Envoy::AccessLog::RequestInfo& request_info) const {
-  return field_extractor_(request_info);
+// TODO(zuercher): If modifying this function to perform header substitutions, consider refactoring
+// this class and RequestHeaderParser to share overlapping code.
+void ResponseHeaderParser::evaluateResponseHeaders(Http::HeaderMap& headers) const {
+  for (const auto& header_value : headers_to_add_) {
+    if (header_value.append_) {
+      headers.addReferenceKey(header_value.header_, header_value.value_);
+    } else {
+      headers.setReferenceKey(header_value.header_, header_value.value_);
+    }
+  }
+
+  for (const auto& header : headers_to_remove_) {
+    headers.remove(header);
+  }
 }
 
 } // namespace Router
